Move custom-gradient-bar into its own folder
