+++ packages/components/src/custom-gradient-picker/utils.js
@@ -9,35 +9,10 @@ import tinycolor from 'tinycolor2';
  */
 import {
 	DEFAULT_GRADIENT,
-	INSERT_POINT_WIDTH,
-	MINIMUM_ABSOLUTE_LEFT_POSITION,
 	HORIZONTAL_GRADIENT_ORIENTATION,
+	DIRECTIONAL_ORIENTATION_ANGLE_MAP,
 } from './constants';
 import { serializeGradient } from './serializer';
-
-export function getHorizontalRelativeGradientPosition(
-	mouseXCoordinate,
-	containerElement,
-	positionedElementWidth
-) {
-	if ( ! containerElement ) {
-		return;
-	}
-	const { x, width } = containerElement.getBoundingClientRect();
-	const absolutePositionValue =
-		mouseXCoordinate -
-		x -
-		MINIMUM_ABSOLUTE_LEFT_POSITION -
-		positionedElementWidth / 2;
-	const availableWidth =
-		width - MINIMUM_ABSOLUTE_LEFT_POSITION - INSERT_POINT_WIDTH;
-	return Math.round(
-		Math.min(
-			Math.max( ( absolutePositionValue * 100 ) / availableWidth, 0 ),
-			100
-		)
-	);
-}
 
 export function getLinearGradientRepresentationOfARadial( gradientAST ) {
 	return serializeGradient( {
@@ -46,21 +21,6 @@ export function getLinearGradientRepresentationOfARadial( gradientAST ) {
 		colorStops: gradientAST.colorStops,
 	} );
 }
-
-const DIRECTIONAL_ORIENTATION_ANGLE_MAP = {
-	top: 0,
-	'top right': 45,
-	'right top': 45,
-	right: 90,
-	'right bottom': 135,
-	'bottom right': 135,
-	bottom: 180,
-	'bottom left': 225,
-	'left bottom': 225,
-	left: 270,
-	'top left': 315,
-	'left top': 315,
-};
 
 function hasUnsupportedLength( item ) {
 	return item.length === undefined || item.length.type !== '%';
